Update README.md
This project aims to implement various genetic algorithms to find optimal or near-optimal solutions for the Traveling Salesman Problem (TSP). The TSP is a well-known combinatorial optimization problem, where the objective is to find the shortest possible route that visits each city exactly once and returns to the starting city.
The genetic algorithms employed in this project will evolve populations of candidate solutions through the processes of selection, crossover, and mutation. These algorithms will be evaluated on TSP instances with 42, 512, and 662 cities, and the distances between cities will be stored in datasets.
Additionally, the project will compare the performance of different genetic algorithms and the time taken by each algorithm to suggest a tour for the given TSP instances.